+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.2.0 (not released)
 - Grad widget
+- Min constrain (only for vertical direction)
 
 ## v0.1.1
 - Fix readme and docs not containing correct links